Added "nextToFront" option, updated CSS to support
CSS updates were only on the compact-theme, everything else didn't need any modification.

Instead of applying border radii to .prev and .next, I simply made it on the first and last list element.
This is because .next may not be at the end anymore.

With this new styling, it also supports the case that .prev or .next DON'T exist, for instance
when the user explicitly removes them (by setting prevText or nextText to false/"" on initialisation).

Regarding CSS browser compatibility, :last-child may not be supported on older browsers, however, it
definitely will be supported by any browser that supports border radius. i.e. if :last-child is NOT supported
then neither is border radius, and hence we're not losing anything.

Considering the use of :last-child, compact-theme took a shift from having border-right: none; to
border-left: none. Now instead of having the border set explicitly for the last child to account for that
missing edge, it's now on the first child (which has wider CSS support, down to IE7) and appropriately border-left.